Move Proto to own module
